Introduce weapon contracts and refactor usages
Add a Contracts namespace and extract weapon-facing interfaces (IWeaponOwnerContext, IWeaponDamageRelay, IWeaponFxRelay, IWeaponFacade, IWeaponManagerFacade, IWeaponViewmodelDriver, IWeaponDataRuntime, IWeaponRuntimeContext). Move IWeaponOwnerContext file to Contracts and add IWeaponViewmodelDriver/related contract files. Refactor multiple classes to depend on these interfaces (PlayerController, Weapon, WeaponManager, WeaponData, WeaponDamageRelay, WeaponFxRelay, Kin* classes, WeaponMount, etc.), add a ResolveViewmodelDriver helper to locate viewmodel drivers by interface, and make small API adjustments (casts, byte enum handling, instance method calls). Purpose: decouple weapon implementations from concrete types for modularity and easier substitution/testing.

+++ Assets/Scripts/Game/Weapon/Core/WeaponCombatAuthority.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Diagnostics;
+using Game.Weapon.Contracts;
 using Game.Weapon.Manager;
 using Network.AntiCheat;
 using Network.Core;
@@ -259,7 +260,7 @@ namespace Game.Weapon.Core {
                 return;
             }
 
-            weaponManager.UpdateServerWeaponState(weaponIndex, reason, localAmmoAfterEvent);
+            weaponManager.UpdateServerWeaponState(weaponIndex, (byte)reason, localAmmoAfterEvent);
         }
 
         [Rpc(SendTo.Server, InvokePermission = RpcInvokePermission.Everyone)]
